Refatoracao do código para aplicar padrões de design
